Global optimization of f(x; a, b)
A wrongheaded effort to trace out the global feasibility set. For the cost function 2a +  3b, objective values at the 3rd relaxation is  -36, 5th is - 30, and 6th is -20. The 6th takes 15 minutes as an SDP to run, and the resultant moment matrices have a zero sparsity pattern. I wonder if there is additional sparsity that can be exploited if not done already.

SparsePOP runs quickly through YALMIP, but I can't figure out  how to increase moment order.