Starting work on JSON, and failed dualization
Dualizing TSSOS is an unnecessary nightmare (24 Gb of memory on this stupid program). Instead, be smart. TSSOS is a moment program, so find the equilvalent SOS program by strong duality. This is done by formulas 22 and 23 in the TSSOS paper.

Export the fblocks and gblocks to json. Use some Yalmip code to turn each block into a polynomial. Form the resultant sos program in yalmip with the decomposed Psatz objective, export to sedumi, and then run decomposed structured subsets.

For tomorrow/future. Done for tonight.